requirements.txt
requests
pandas
openpyxl
pyyaml
tqdm
tenacity
python-dotenv
